[#97] Refactor K8S API
