+++ legion/legion/k8s/__init__.py
@@ -16,10 +16,11 @@
 """
 legion k8s module
 """
-from legion.k8s.enclave import Enclave
+from legion.k8s.enclave import Enclave, find_enclaves
 from legion.k8s.services import Service, ModelService
 from legion.k8s.definitions import ModelIdVersion, ModelDeploymentDescription
 from legion.k8s.definitions import STATUS_OK, STATUS_FAIL, STATUS_WARN
 from legion.k8s.definitions import EVENT_ADDED, EVENT_MODIFIED, EVENT_DELETED
-from legion.k8s.utils import build_client, get_current_namespace, find_enclaves, load_config, load_secrets
+from legion.k8s.utils import build_client, get_current_namespace, load_config, load_secrets
 from legion.k8s.utils import CONNECTION_CONTEXT
+from legion.k8s.watch import ResourceWatch
