+++ legion/legion/k8s/utils.py
@@ -40,114 +40,10 @@ import legion.k8s.services
 import legion.k8s.enclave
 from legion.k8s.definitions import \
     LEGION_COMPONENT_LABEL, LEGION_COMPONENT_NAME_MODEL, \
-    ENCLAVE_NAMESPACE_LABEL, LEGION_SYSTEM_LABEL, LEGION_SYSTEM_VALUE
+    LEGION_SYSTEM_LABEL, LEGION_SYSTEM_VALUE
 
 LOGGER = logging.getLogger(__name__)
 CONNECTION_CONTEXT = None
-
-
-class ResourceWatch:
-    """
-    Watch for K8S resources (context manager)
-    """
-
-    def __init__(self, api_function, *args,
-                 filter_callable=None,
-                 object_constructor=None,
-                 **kwargs):
-        """
-        Initialize context manager for resource watch
-
-        :param api_function: API function to watch
-        :param args: additional positional arguments for API function
-        :param filter_callable: (Optional) callable to filter objects
-        :param object_constructor:  (Optional) callable to construct object wrappers
-        :param kwargs: additional key value arguments for API function
-        """
-        self._api_function = api_function
-        self._filter_callable = filter_callable
-        self._object_constructor = object_constructor
-        self._args = args
-        self._kwargs = kwargs
-
-        self._watch = kubernetes.watch.Watch()
-        self._stream = None
-
-    def __enter__(self):
-        """
-        Enter watch
-
-        :return:
-        """
-        self._stream = self._watch.stream(self._api_function,
-                                          *self._args,
-                                          **self._kwargs)
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        """
-        Exit watch
-
-        :param exc_type: exception type
-        :param exc_val: exception
-        :param exc_tb: exception traceback
-        :return: None
-        """
-        self.stop()
-
-    @property
-    def stream(self):
-        """
-        Access watch stream
-
-        :return: tuple(str, Any) -- event type and event object (plain or constructed if
-                                    object_constructor has been passed)
-        """
-        LOGGER.debug('Starting watch stream')
-
-        reconnect = True
-        while reconnect:
-            reconnect = False
-
-            try:
-                for event in self._stream:
-                    event_type = event['type']
-                    event_object = event['object']
-
-                    # Check if valid object
-                    pass_event = not self._filter_callable or self._filter_callable(event_object)
-
-                    if pass_event:
-                        # Construct specific object
-                        if self._object_constructor:
-                            event_object = self._object_constructor(event_object)
-
-                        yield (event_type, event_object)
-            except urllib3.exceptions.ProtocolError:
-                LOGGER.info('Connection to K8S API has been lost, reconnecting..')
-                reconnect = True
-
-        LOGGER.debug('Watch stream has been ended')
-
-    @property
-    def watch(self):
-        """
-        Access watch object
-
-        :return: :py:class:`kubernetes.watch.Watch` -- watch object
-        """
-        return self._watch
-
-    def stop(self):
-        """
-        Stop watch
-
-        :return: None
-        """
-        try:
-            self.watch.stop()
-        finally:
-            LOGGER.debug('Watch has been stopped')
 
 
 def build_client():
@@ -231,168 +127,6 @@ def load_secrets(path_to_secrets):
     return secrets
 
 
-def find_model_deployment(model_id, namespace='default'):
-    """
-    Find model deployment by model id
-
-    :param model_id: model id
-    :type model_id: str
-    :param namespace: namespace
-    :type namespace: str
-    :return: :py:class:`kubernetes.client.models.extensions_v1beta1_deployment.ExtensionsV1beta1Deployment`
-    """
-    client = build_client()
-
-    extension_api = kubernetes.client.ExtensionsV1beta1Api(client)
-    all_deployments = extension_api.list_namespaced_deployment(namespace)
-
-    type_label_name = normalize_name(legion.containers.headers.DOMAIN_CONTAINER_TYPE)
-    type_label_value = 'model'
-
-    model_id_name = normalize_name(legion.containers.headers.DOMAIN_MODEL_ID)
-    model_id_value = normalize_name(model_id)
-
-    for deployment in all_deployments.items:
-        if deployment.metadata.labels.get(type_label_name) == type_label_value \
-                and deployment.metadata.labels.get(model_id_name) == model_id_value:
-            return deployment
-
-    return None
-
-
-def find_all_models_deployments(namespace='default'):
-    """
-    Find all models deployments
-
-    :param namespace: namespace
-    :type namespace: str or none for all
-    :return: list[:py:class:`kubernetes.client.models.extensions_v1beta1_deployment.ExtensionsV1beta1Deployment`]
-    """
-    client = build_client()
-
-    extension_api = kubernetes.client.ExtensionsV1beta1Api(client)
-    if namespace:
-        all_deployments = extension_api.list_namespaced_deployment(namespace)
-    else:
-        all_deployments = extension_api.list_deployment_for_all_namespaces()
-
-    type_label_name = normalize_name(legion.containers.headers.DOMAIN_CONTAINER_TYPE)
-    type_label_value = 'model'
-
-    model_deployments = [
-        deployment
-        for deployment in all_deployments.items
-        if deployment.metadata.labels.get(type_label_name) == type_label_value
-    ]
-
-    return model_deployments
-
-
-def find_all_services(namespace='', component=''):
-    """
-    Find all services details by criteria
-
-    :param namespace: namespace
-    :type namespace: str or none for all
-    :param component: filter by specified component value, or none for all
-    :type component: str
-    :return: list[V1Service]
-    """
-    client = build_client()
-
-    core_api = kubernetes.client.CoreV1Api(client)
-    if namespace:
-        all_services = core_api.list_namespaced_service(namespace)
-    else:
-        all_services = core_api.list_service_for_all_namespaces()
-
-    if component:
-        all_services = [
-            service
-            for service in all_services.items
-            if service.metadata.labels.get(LEGION_COMPONENT_LABEL) == component
-        ]
-    else:
-        all_services = all_services.items
-
-    return all_services
-
-
-def get_service(namespace='', component=''):
-    """
-    Get a service details by criteria
-
-    :param namespace: namespace
-    :type namespace: str or none
-    :param component: filter by specified component value, or none for all
-    :type component: str
-    :return: V1Service or None
-    """
-    result = find_all_services(namespace, component)
-    return legion.k8s.services.Service(result[0]) if result else None
-
-
-def find_all_ingresses(namespace='', component=''):
-    """
-    Find all ingresses details by criteria
-
-    :param namespace: namespace
-    :type namespace: str or none for all
-    :type component: filter by specified component value, or none for all
-    :return: list[V1beta1Ingress]
-    """
-    client = build_client()
-
-    extension_api = kubernetes.client.ExtensionsV1beta1Api(client)
-    if namespace:
-        all_ingresses = extension_api.list_namespaced_ingress(namespace)
-    else:
-        all_ingresses = extension_api.list_ingress_for_all_namespaces()
-
-    if component:
-        all_ingresses = [
-            ingress
-            for ingress in all_ingresses.items
-            if ingress.metadata.labels.get(LEGION_COMPONENT_LABEL) == component
-        ]
-    else:
-        all_ingresses = all_ingresses.items
-
-    return all_ingresses
-
-
-def get_ingress(namespace='', component=''):
-    """
-    Get ingress details by criteria
-
-    :param namespace: namespace
-    :type namespace: str or none for all
-    :type component: filter by specified component value, or none for all
-    :return: V1beta1Ingress -- Ingress object or None
-    """
-    ingresses = find_all_ingresses(namespace, component)
-    return ingresses[0] if ingresses else None
-
-
-def get_ingress_url(ingress):
-    """
-    Get URL of ingress object
-
-    :param ingress: ingress object
-    :type ingress: V1beta1Ingress
-    :return: str or None -- URL from tls section,
-    """
-    spec = ingress.spec
-
-    host = spec.rules[0].host if spec.rules else None
-    if not host:
-        return None
-
-    protocol = 'https' if spec.tls and any(host in tls.hosts for tls in spec.tls) else 'http'
-
-    return '{}://{}'.format(protocol, host)
-
-
 def get_docker_image_labels(image):
     """
     Get labels from docker image
@@ -451,22 +185,3 @@ def get_meta_from_docker_labels(labels):
     compatible_labels[LEGION_SYSTEM_LABEL] = LEGION_SYSTEM_VALUE
 
     return normalize_name(k8s_name, dns_1035=True), compatible_labels, model_id, model_version
-
-
-def find_enclaves():
-    """
-    Get a list of enclaves
-
-    :return list[Enclave] -- list of found enclaves
-    """
-    client = build_client()
-
-    core_api = kubernetes.client.CoreV1Api(client)
-    all_namespaces = core_api.list_namespace()
-
-    enclaves = []
-    for namespace in all_namespaces.items:
-        if namespace.metadata.labels and ENCLAVE_NAMESPACE_LABEL in namespace.metadata.labels:
-            enclaves.append(legion.k8s.enclave.Enclave(namespace.metadata.name))
-
-    return enclaves
